docs: overhaul FCU hardware, notch, and audit notes

+++ lib/DShotRMT/src/dshot_utils.h
@@ -125,7 +125,9 @@ inline void printDShotResult(dshot_result_t &result, Stream &output = Serial)
     // Print telemetry data if available
     if (result.success && (result.erpm > 0 || result.motor_rpm > 0))
     {
-        output.printf(" | eRPM: %u, Motor RPM: %u", result.erpm, result.motor_rpm);
+        output.printf(" | eRPM: %lu, Motor RPM: %lu",
+                      static_cast<unsigned long>(result.erpm),
+                      static_cast<unsigned long>(result.motor_rpm));
     }
 
     output.println();
@@ -159,7 +161,9 @@ inline void printDShotInfo(DShotRMT &dshot_rmt, Stream &output = Serial)
 
             if (telemetry_result.erpm > 0 || telemetry_result.motor_rpm > 0)
             {
-                output.printf("  eRPM: %u, Motor RPM: %u\n", telemetry_result.erpm, telemetry_result.motor_rpm);
+                output.printf("  eRPM: %lu, Motor RPM: %lu\n",
+                              static_cast<unsigned long>(telemetry_result.erpm),
+                              static_cast<unsigned long>(telemetry_result.motor_rpm));
             }
 
             if (telemetry_result.telemetry_available)
